Add similarity  comparision
